+++ tap_mambu/helpers/client.py
@@ -1,9 +1,11 @@
 import backoff
 import requests
 import requests.adapters
-from requests.exceptions import ConnectionError
+from requests.exceptions import ConnectionError, ChunkedEncodingError
 from singer import metrics, get_logger
 
+from urllib3.exceptions import ProtocolError
+from tap_mambu.helpers.constants import DEFAULT_DATE_WINDOW_SIZE
 
 LOGGER = get_logger()
 class ClientError(Exception):
@@ -118,13 +120,21 @@ class MambuClient(object):
                  subdomain,
                  apikey_audit,
                  page_size,
-                 user_agent=''):
+                 user_agent='',
+                 window_size=DEFAULT_DATE_WINDOW_SIZE):
         self.__username = username
         self.__password = password
         self.__subdomain = subdomain
         base_url = "https://{}.mambu.com/api".format(subdomain)
         self.base_url = base_url
         self.page_size = page_size
+        try:
+            self.window_size = int(float(window_size)) if window_size else DEFAULT_DATE_WINDOW_SIZE
+            if self.window_size <= 0:
+                raise ValueError()
+        except ValueError:
+            raise Exception("The entered window size '{}' is invalid; it should be a valid non-zero integer.".format(window_size))
+
         self.__user_agent = f'MambuTap-{user_agent}' if user_agent else 'MambuTap'
         self.__apikey = apikey
         self.__session = requests.Session()
@@ -178,7 +188,8 @@ class MambuClient(object):
             return True
 
     @backoff.on_exception(backoff.expo,
-                          (MambuInternalServiceError, ConnectionError, MambuApiLimitError),
+                          (MambuInternalServiceError, ConnectionError,
+                           ChunkedEncodingError, MambuApiLimitError, ProtocolError),
                           max_tries=7,
                           factor=3)
     def request(self, method, path=None, url=None, json=None, version=None, apikey_type=None, **kwargs):