+++ tap_mambu/tap_generators/child_generator.py
@@ -1,4 +1,5 @@
 from .generator import TapGenerator
+from typing import List
 
 
 class ChildGenerator(TapGenerator):
@@ -9,4 +10,8 @@ class ChildGenerator(TapGenerator):
     def _init_endpoint_config(self):
         super(ChildGenerator, self)._init_endpoint_config()
         self.endpoint_path = f"{self.endpoint_parent_id}"  # include parent id in endpoint path
+
+    def _init_buffers(self):
+        self.buffer: List = list()
+        self.max_buffer_size = 1000