+++ tests/base.py
@@ -249,7 +249,8 @@ class MambuBaseTest(unittest.TestCase):
             'start_date': '2017-01-01T00:00:00Z',
             'username': os.environ['TAP_MAMBU_USERNAME'],
             'subdomain': os.environ['TAP_MAMBU_SUBDOMAIN'],
-            'page_size': '100'
+            'page_size': '90',
+            'window_size': 30
         }
 
         if not original_properties: